+++ src/pages/jinwookim/Login/Login.js
@@ -1,4 +1,3 @@
-/* eslint-disable jsx-a11y/anchor-is-valid */
 import React from "react";
 import "./Login.scss";
 import { useNavigate } from "react-router-dom";
